% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acf_acf.R
\name{Acf_acf}
\alias{Acf_acf}
\title{Autocorrelation wrapper function to `apply` over xts}
\usage{
Acf_acf(x)
}
\arguments{
\item{x}{Vector of ordered observations.}
}
\description{
Autocorrelation wrapper function to `apply` over xts
}
