% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fisher.R
\name{Fisher}
\alias{Fisher}
\title{Fisher Transform Nominal time series to Real}
\usage{
Fisher(Nominal, price, quantity, baseYear = 249)
}
\arguments{
\item{Nominal}{xts object containing the names of the prices index.}

\item{price}{Vector that containing the prices index.}

\item{quantity}{Vector that contains the quantities index.}

\item{baseYear}{The base period(s) to calculate the indices, corresponding to
the row, or observation number of the series.
baseYear default period is 250 for Q2 2009.}
}
\description{
Wrapper function calculates Fisher price index and applies it to an xts object
}
\details{
\code{Fisher(Nominal, price, quantity, baseYear=250)}
}
\examples{
\dontrun{
Real_PCE <- Fisher(PCE, PCE_Price, PCE_Quanity, 249)
}

}
