% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HL_filter.R
\name{HL_filter}
\alias{HL_filter}
\title{Hamilton Filtered quarterly xts object.}
\usage{
HL_filter(x, h = 8, ...)
}
\arguments{
\item{x}{An xts object of quarterly periodicity.}

\item{h}{The look ahead parameter indicating the length of the AR(4) lag. Default to h = 8, or 8 quarters.}

\item{...}{see "lm"}
}
\description{
\code{HL_filter} returns an xts object containing the original series,
                 the difference between t+h and t, the fitted (trend) values, as well
                 as the residual (cyclical) component.
}
\examples{
\dontrun{
HL_filter(PAYEMS, h = 8)
}

}
\references{
James D. Hamilton. Why You Should Never Use the Hodrick-Prescott Filter.
           NBER Working Paper No. 23429, Issued in May 2017.
}
