% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HL_filter_ar.R
\name{HL_filter_ar}
\alias{HL_filter_ar}
\title{Hamilton Filtered ar(x) quarterly xts object.}
\usage{
HL_filter_ar(x, h = 8, p = 4, ...)
}
\arguments{
\item{x}{An xts object of quarterly periodicity.}

\item{h}{The look ahead parameter indicating the length of the first lag. Default to h = 8, or 8 quarters.}

\item{p}{The number of lagged AR terms. Default is p = 4, to reflect 4 quarters of an annual time series.}

\item{...}{see "lm"}
}
\description{
\code{HL_filter_ar} returns an xts object containing the original series,
                 the difference between t+h and t, the fitted (trend) values, as well
                 as the residual (cyclical) component.
}
\examples{
\dontrun{
HL_filter(PAYEMS, h = 8)
}

}
\references{
James D. Hamilton. Why You Should Never Use the Hodrick-Prescott Filter.
           NBER Working Paper No. 23429, Issued in May 2017.
}
