% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HL_filter_lm.R
\name{HL_filter_lm}
\alias{HL_filter_lm}
\title{Hamilton Filtered quarterly time series Linear Model}
\usage{
HL_filter_lm(x, h = 8, ...)
}
\arguments{
\item{x}{An xts object of quarterly periodicity.}

\item{h}{The look ahead parameter indicating the length of the AR(4) lag. Default to h = 8, or 8 quarters.}

\item{...}{see "lm"}
}
\description{
For time series of quarterly periodicity, Hamilton's solution can be described as an \eqn{AR(4)} process, 
additionally lagged by \eqn{h} periods. \deqn{y_{t+h} = \beta_0 + \beta_1 y_t + \beta_2 y_{t-1} +\beta_3 y_{t-2} + \beta_4 y_{t-3} + v_{t+h}}
\deqn{\hat{v}_{t+h} = y_{t+h} + \hat{\beta}_0 + \hat{\beta}_1 y_t + \hat{\beta}_2 y_{t-1} + \hat{\beta}_3 y_{t-2} + \hat{\beta}_4 y_{t-3}}
}
\details{
\value{
\strong\code{HL_filter_lm} returns a linear model of class "formula", like the output created by `stats::lm`.
}
}
\examples{
HL_filter_lm(GDPC1, h = 8)

}
\references{
James D. Hamilton. Why You Should Never Use the Hodrick-Prescott Filter.
           NBER Working Paper No. 23429, Issued in May 2017.
}
