% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpfilter.R
\name{hpfilter}
\alias{hpfilter}
\title{Hodrick-Prescott filter of a time series}
\usage{
hpfilter(x, lambda = 1600)
}
\arguments{
\item{x}{A regular time series.}

\item{lambda}{The smoothing parameter (lambda) of the Hodrick-Prescott filter}
}
\description{
\code{hpfilter()} This function implements the Hodrick-Prescott for estimating
                 cyclical and trend component of a time series. The function
                 computes cyclical and trend components of the time series
                 using a frequency cut-off or smoothness parameter.
}
\examples{
\dontrun{
hp_PAYEMS <- hpfilter(PAYEMS, lambda = 1600)
}
}
\references{
R.J. Hodrick and E.C. Prescott. Postwar US business cycles: an empirical investigation. Journal of Money, Credit, and Banking, 29(1):1-16, 1997.
}
