% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPGSC1.R
\docType{data}
\name{IMPGSC1}
\alias{IMPGSC1}
\title{Real imports of goods and services}
\format{An \code{\link{xts}} object containing quarterly observations of Real imports of goods and services.
 \itemize{
 \item\strong{Release:} {Gross Domestic Product}
 \item\strong{Seasonal Adjustment:} {Seasonally Adjusted Annual Rate} 
 \item\strong{Frequency:} {Quarterly}
 \item\strong{Units:} {Billions of Chained 2009 Dollars}
 \item\strong{Date Range:} {1947-01-01 to 2017-07-01}
 \item\strong{Last Updated} {2017-12-21 7:51 AM CST}
}}
\source{
U.S. Bureau of Economic Analysis \url{https://fred.stlouisfed.org/data/IMPGSC1.txt}
}
\usage{
data(IMPGSC1)
}
\description{
\code{IMPGSC1} Real imports of goods and services
}
\section{Notes}{

BEA Account Code: A021RX
A Guide to the National Income and Product Accounts of the United States (NIPA) 
- (http://www.bea.gov/national/pdf/nipaguid.pdf)
}

\examples{
data(IMPGSC1)
tail(IMPGSC1)
plot(IMPGSC1,grid.col = "white", col="blue")
}
\keyword{datasets}
