% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yth_cycle.R
\name{yth_cycle}
\alias{yth_cycle}
\title{yth_cycle}
\usage{
yth_cycle(x, h = 8, p = 4, ...)
}
\arguments{
\item{x}{An xts object.}

\item{h}{The look ahead parameter. Default to h = 8, or 8 quarters for two years.}

\item{p}{Idicating the number of lags. Default to p = 4, or 4 quarters for one year.}

\item{...}{additional argeuments of accepted by the model class see "lm"}
}
\value{
\code{yth_cycle} returns a xts object of the cyclical component, the residual of the yt_ar model.
}
\description{
An alternative to the Hodrick-Prescott Filter
}
\details{
For time series of quarterly periodicity, Hamilton suggests an \eqn{AR(4)} process,
additionally lagged by \eqn{h} periods. \deqn{y_{t+h} = \beta_0 + \beta_1 y_t + \beta_2 y_{t-1} + \beta_3 y_{t-2} + \beta_4 y_{t-3} + v_{t+h}}
\deqn{\hat{v}_{t+h} = y_{t+h} - \hat{\beta}_0 + \hat{\beta}_1 y_t + \hat{\beta}_2 y_{t-1} + \hat{\beta}_3 y_{t-2} + \hat{\beta}_4 y_{t-3}}
}
\examples{
yth_cycle(GDPC1, h = 8, p = 4)

}
\references{
James D. Hamilton. "Why You Should Never Use the Hodrick-Prescott Filter".
           NBER Working Paper No. 23429, Issued in May 2017.
}
