% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDPC1.R
\docType{data}
\name{GDPC1}
\alias{GDPC1}
\title{GDPC1}
\format{An xts object containing quarterly observations of Real GDP
\itemize{
 \item \strong {GDPC1} Real Gross Domestic Product,  Seasonally Adjusted Annual Rate, in Billions of Chained 2009 Dollars.
}}
\source{
U.S. Bureau of Economic Analysis \url{https://fred.stlouisfed.org/data/GDPC1.txt}
}
\usage{
data("GDPC1")
}
\description{
Real gross domestic product is the inflation adjusted value of the
goods and services produced by labor and property located in the
United States.For more information see the Guide to the National
Income and Product Accounts of the United States (NIPA) -
(http://www.bea.gov/national/pdf/nipaguid.pdf)
}
\section{Notes}{

}

\examples{
 tail(GDPC1)
}
\keyword{datasets}
