% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDIC1.R
\docType{data}
\name{GPDIC1}
\alias{GPDIC1}
\title{Real Gross Private Domestic Investment}
\format{
An \code{\link[xts]{xts}} object containing quarterly observations of
Real Gross Private Domestic Investment.
\describe{
  \item{**Release:**}{Gross Domestic Product}
  \item{**Seasonal Adjustment:**}{Seasonally Adjusted Annual Rate}
  \item{**Frequency:**}{Quarterly}
  \item{**Units:**}{Billions of Chained 2012 Dollars}
  \item{**Date Range:**}{1947-01-01 to 2021-07-01}
  \item{**Last Updated:**}{2021-12-22 7:53 AM CST}
}
}
\source{
U.S. Bureau of Economic Analysis \url{https://fred.stlouisfed.org/data/GPDIC1.txt}
}
\usage{
data(GPDIC1)
}
\description{
\code{GPDIC1} Real Gross Private Domestic Investment
}
\section{Notes}{

BEA Account Code: A006RX

A Guide to the National Income and Product Accounts of the United States (NIPA)
- (http://www.bea.gov/national/pdf/nipaguid.pdf)
}

\examples{
data(GPDIC1)
tail(GPDIC1)
plot(GPDIC1, grid.col = "white", col = "blue")
}
\keyword{datasets}
