% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDIC1.R
\docType{data}
\name{GPDIC1}
\alias{GPDIC1}
\title{Real Gross Private Domestic Investment}
\format{An \code{\link{xts}} object of Real Exports of Goods and Services.
 \itemize{
 \item\strong{Release:} {Gross Domestic Product}
 \item\strong{Seasonal Adjustment:} {Seasonally Adjusted Annual Rate} 
 \item\strong{Frequency:} {Quarterly}
 \item\strong{Units:} {Billions of Chained 2012 Dollars}
 \item\strong{Date Range:} {1947-01-01 to 2019-04-01}
 \item\strong{Last Updated} { 2019-08-29 7:59 AM CDT}
}}
\source{
U.S. Bureau of Economic Analysis \url{https://fred.stlouisfed.org/data/GPDIC1.txt}
}
\usage{
data(GPDIC1)
}
\description{
\code{GPDIC1} Real Gross Private Domestic Investment
}
\section{Notes}{

BEA Account Code: A006RX

A Guide to the National  Income and Product Accounts of the United States (NIPA) 
- (http://www.bea.gov/national/pdf/nipaguid.pdf)
}

\examples{
data(GPDIC1)

tail(GPDIC1)

plot(GPDIC1, grid.col = "white", col="blue")
}
\keyword{datasets}
