% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HL_filter_lm.R
\name{HL_filter_lm}
\alias{HL_filter_lm}
\title{Linear Model Hamilton Filtered quarterly xts object.}
\usage{
HL_filter_lm(x, h = 8, ...)
}
\arguments{
\item{x}{An xts object of quarterly periodicity.}

\item{h}{The look ahead parameter indicating the length of the AR(4) lag. Default to h = 8, or 8 quarters.}

\item{...}{see "lm"}
}
\description{
\code{HL_filter_lm} returns a hamilton filtered linear model of class "formula"
}
\examples{
\dontrun{
HL_filter_lm(GDPC1, h = 8)
}
}
\references{
James D. Hamilton. Why You Should Never Use the Hodrick-Prescott Filter.
           NBER Working Paper No. 23429, Issued in May 2017.
}
