% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPGSC1.R
\docType{data}
\name{IMPGSC1}
\alias{IMPGSC1}
\title{Real Imports of Goods and Services}
\format{
An \code{\link[xts]{xts}} object containing quarterly observations of
Real Imports of Goods and Services.
\describe{
  \item{**Release:**}{Gross Domestic Product}
  \item{**Seasonal Adjustment:**}{Seasonally Adjusted Annual Rate}
  \item{**Frequency:**}{Quarterly}
  \item{**Units:**}{Billions of Chained 2012 Dollars}
  \item{**Date Range:**}{1947-01-01 to 2021-07-01}
  \item{**Last Updated:**}{2021-12-22 7:53 AM CST}
}
}
\source{
U.S. Bureau of Economic Analysis \url{https://fred.stlouisfed.org/data/IMPGSC1.txt}
}
\usage{
data(IMPGSC1)
}
\description{
\code{IMPGSC1} Real Imports of Goods and Services
}
\section{Notes}{

BEA Account Code: A021RX

A Guide to the National Income and Product Accounts of the United States (NIPA)
- (http://www.bea.gov/national/pdf/nipaguid.pdf)
}

\examples{
data(IMPGSC1)
tail(IMPGSC1)
plot(IMPGSC1, grid.col = "white", col = "blue")
}
\keyword{datasets}
