% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yth_filter.R
\name{yth_filter}
\alias{yth_filter}
\title{Filtered xts object}
\usage{
yth_filter(x, h = 8, p = 4, output = c("x", "trend", "cycle", "random"),
  ...)
}
\arguments{
\item{x}{A univariate \code{\link{xts}} object of any \code{\link{zoo}} index class,
such as \code{\link{Date}}, \code{\link{yearmon}}, or \code{\link{yearqtr}}.
For converting objects of type \code{timeSeries, ts, irts, fts, matrix, data.frame}
or \code{zoo} to \code{\link{xts}}, please read \code{\link{as.xts}}.}

\item{h}{An \code{\link{integer}}, defining the lookahead period.
Defaults to \code{h = 8}. The default assumes economic data of quarterly 
periodicity with a lookahead period of 2 years. This function is not limited 
by the default parameter, and Econometricians may change it as required.}

\item{p}{An \code{\link{integer}}, indicating the number of lags. A Default of 
\code{p = 4}, assumes data is of quarterly periodicity. If data is of monthly 
periodicity, one may choose \code{p = 12} or aggregate the series to quarterly 
periodicity and maintain the default. Econometricians should use this parameter 
to accommodate the Seasonality of their data.}

\item{output}{A character vector determining the output of this function. 
Defaults to \code{output = c("x","trend", "cycle", "random")}, which returns the original time series 
(\code{x}), yth_glm \code{\link{fitted.values}} (\code{"trend"}), yth_glm \code{\link{residuals}}
(\code{"cycle"}), and a random walk series defined by differencing \eqn{y_{t+h}} 
and \eqn{y_t} (\code{"random"}). Arguments \code{"x"}, \code{"trend"}, \code{"cycle"}, and 
\code{"random"} extract their corresponding univariate series and can be merged in any combination.
For example \code{c("x", "trend")} returns both the original series "x" and the "trend" components. 
\code{c("cycle", "random")} will return both the "cycle" and "random" components.}

\item{...}{other arguments passed to the function \code{\link[stats]{glm}}}
}
\value{
If the default parameter of output = "All" is used, the function returns 
 an \code{\link{xts}} object containing the original time series (x), 
 \code{\link{yth_glm}} model \code{\link{fitted.values}},  \code{\link{yth_glm}} model 
 \code{\link{residuals}}, and a random walk series defined by differencing 
 \eqn{y_{t+h}} and \eqn{y_t}. Other output choices lead to various combinations 
 of these time series.
}
\description{
\code{yth_filter} returns an \code{\link{xts}} object containing user defined
 combinations of the original, trend, cycle, and random walk series.
}
\details{
For time series of quarterly periodicity, Hamilton suggests parameters of
 h = 8 and p = 4, or an \eqn{AR(4)} process, additionally lagged by \eqn{8}
 lookahead periods. Econometricians may explore variations of h. However, p 
 is designed to correspond with the seasonality of a given periodicity and 
 should be matched accordingly.
 \deqn{y_{t+h} = \beta_0 + \beta_1 y_t + \beta_2 y_{t-1} + \beta_3 y_{t-2} + \beta_4 y_{t-3} + v_{t+h}}
 \deqn{\hat{v}_{t+h} = y_{t+h} - \hat{\beta}_0 + \hat{\beta}_1 y_t + \hat{\beta}_2 y_{t-1} + \hat{\beta}_3 y_{t-2} + \hat{\beta}_4 y_{t-3}}
 Which can be rewritten as:
 \deqn{y_{t} = \beta_0 + \beta_1 y_{t-8} + \beta_2 y_{t-9} + \beta_3 y_{t-10} + \beta_4 y_{t-11} + v_{t}}
 \deqn{\hat{v}_{t} = y_{t} - \hat{\beta}_0 + \hat{\beta}_1 y_{t-8} + \hat{\beta}_2 y_{t-9} + \hat{\beta}_3 y_{t-10} + \hat{\beta}_4 y_{t-11}}
}
\examples{
data(GDPC1)
log_GDP <- 100*log(GDPC1)

GDP_filtered <- yth_filter(log_GDP, h = 8, p = 4)
tail(GDP_filtered, 8)

GDP_yth_cycle <- yth_filter(log_GDP, h = 8, p = 4, output = c("cycle", "random"))
main <- "Cyclical and random component of 100*log(Real GDP)"
plot(GDP_yth_cycle, grid.col = "white", legend.loc = "topright", main = main)

}
\references{
James D. Hamilton. \href{http://econweb.ucsd.edu/~jhamilto/hp.pdf}{Why You Should Never Use the Hodrick-Prescott Filter}.
           NBER Working Paper No. 23429, Issued in May 2017.
}
\seealso{
\code{\link{yth_glm}}
}
