% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_filter.R
\name{yth_filter}
\alias{yth_filter}
\title{Hamilton Filtered quarterly time series}
\usage{
yth_filter(x, h = 8, p = 4, ...)
}
\arguments{
\item{x}{An xts object.}

\item{h}{The look ahead parameter. Default to h = 8, or 8 quarters for two years.}

\item{p}{Idicating the number of lags. Default to p = 4, or 4 quarters for one year.}

\item{...}{see "lm"}
}
\value{
\code{yth_filter} returns an xts object containing the original series,
                  the fitted (trend) values, the residual (cycle) component,
                  and a "random walk" series represented by the difference between \eqn{y_{t+h}} and \eqn{y_t}
}
\description{
For time series of quarterly periodicity, Hamilton's solution can be described as an \eqn{AR(4)} process,
additionally lagged by \eqn{h} periods. \deqn{y_{t+h} = \beta_0 + \beta_1 y_t + \beta_2 y_{t-1} +\beta_3 y_{t-2} + \beta_4 y_{t-3} + v_{t+h}}
\deqn{\hat{v}_{t+h} = y_{t+h} + \hat{\beta}_0 + \hat{\beta}_1 y_t + \hat{\beta}_2 y_{t-1} + \hat{\beta}_3 y_{t-2} + \hat{\beta}_4 y_{t-3}}
}
\examples{
GDPC1_filtered <- yth_filter(GDPC1, h = 8, p = 4)
plot(GDPC1_filtered)

}
\references{
James D. Hamilton. Why You Should Never Use the Hodrick-Prescott Filter.
           NBER Working Paper No. 23429, Issued in May 2017.
}
