% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yth_glm.R
\name{yth_glm}
\alias{yth_glm}
\title{A better alternative to the Hodrick-Prescott Filter}
\usage{
yth_glm(x, h = 8, p = 4, ...)
}
\arguments{
\item{x}{An univariate xts series of zoo index class, such as "Date", "yearmon", or "yearqtr".}

\item{h}{The look ahead parameter. Default to h = 8, or 8 quarters for two years of look ahead data.}

\item{p}{Idicating the number of lags. Default to p = 4, or 4 quarters for one year.}

\item{...}{see "glm"}
}
\value{
\code{yth_glm} returns a linear model estimed by ordinary least squares of class "formula".
}
\description{
For time series of quarterly periodicity, Hamilton suggests an \eqn{AR(4)} process,
 regressed against a look ahead of \eqn{h} periods. \deqn{y_{t+h} = \beta_0 + \beta_1 y_t + \beta_2 y_{t-1} + \beta_3 y_{t-2} + \beta_4 y_{t-3} + v_{t+h}}
 \deqn{\hat{v}_{t+h} = y_{t+h} - \hat{\beta}_0 + \hat{\beta}_1 y_t + \hat{\beta}_2 y_{t-1} + \hat{\beta}_3 y_{t-2} + \hat{\beta}_4 y_{t-3}}
}
\examples{
yth_glm(GDPC1, h = 8, p = 4)

}
\references{
James D. Hamilton. "Why You Should Never Use the Hodrick-Prescott Filter".
           NBER Working Paper No. 23429, Issued in May 2017.
}
